# Conflicts:
#	src/main/java/com/project/foodiefoodie/config/InterceptorConfig.java
#	src/main/java/com/project/foodiefoodie/member/domain/Master.java
#	src/main/java/com/project/foodiefoodie/member/repository/MasterMapper.java
#	src/main/java/com/project/foodiefoodie/member/service/MasterService.java
#	src/main/java/com/project/foodiefoodie/proBoard/controller/ProBoardController.java
#	src/main/java/com/project/foodiefoodie/proBoard/service/ProBoardService.java
#	src/main/resources/com/project/foodiefoodie/mainpage/repository/MainPageMapper.xml
#	src/main/resources/com/project/foodiefoodie/member/repository/MasterMapper.xml
#	src/main/resources/com/project/foodiefoodie/proBoard/repository/ProBoardMapper.xml

+++ src/main/java/com/project/foodiefoodie/review/dto/AvgStarDTO.java
@@ -11,5 +11,5 @@ public class AvgStarDTO {
 
     private String businessNo;
     private int reviewCount;
-    private double sr;
+    private int sr;
 }
